# accident/infra/api/openapi.yaml
---
paths:
  '/sickLeaves/{id}/requestsickleavebenefit':
    summary: request sickLeave benefit operation on sickLeaves
    put:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                employeeId:
                  description: employeeId of this RequestSickLeaveBenefit
                  type: string
                businessCode:
                  description: businessCode of this RequestSickLeaveBenefit
                  type: string
                period:
                  description: period of this RequestSickLeaveBenefit
                  type: integer
                id:
                  description: id of this RequestSickLeaveBenefit
                  type: integer
            examples:
              Example 0:
                value:
                  employeeId: user01
                  businessCode: bc_1
                  period: 10
      parameters:
        - name: id
          description: RequestSickLeaveBenefit name
          schema:
            type: integer
          in: path
          required: true
          examples:
            Example 0:
              value: 1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SickLeave'
              examples:
                Example 0:
                  value:
                    id: 1
                    accessmentId: 1
                    accidentId: 1
                    businessCode: bc_1
                    employeeId: user01
                    averageSalary: null
                    period: 10
                    status: 휴업급여 요청됨
                    date: N/A
          description: RequestSickLeaveBenefit Operation
      operationId: RequestSickLeaveBenefit
      summary: RequestSickLeaveBenefit
      description: RequestSickLeaveBenefit
  /sickLeaves:
    summary: Global operations on sickLeaves
    get:
      tags:
        - sickLeaves
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SickLeave'
              examples:
                sickLeaves_json:
                  value:
                    - id: 1
                      accessmentId: 1
                      accidentId: 1
                      businessCode: A string
                      employeeId: A string
                      averageSalary: value
                      period: 1
                      status: A string
                      date: value
          description: Get list of sickLeaves
      operationId: GetSickLeave
      summary: Get list of sickLeaves
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SickLeave'
            examples:
              '1':
                value:
                  id: 1
                  accessmentId: 1
                  accidentId: 1
                  businessCode: A string
                  employeeId: A string
                  averageSalary: value
                  period: 1
                  status: A string
                  date: value
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SickLeave'
              examples:
                '1':
                  value:
                    id: 1
                    accessmentId: 1
                    accidentId: 1
                    businessCode: A string
                    employeeId: A string
                    averageSalary: value
                    period: 1
                    status: A string
                    date: value
          description: Changed SickLeave
      operationId: Post SickLeave
      summary: Post SickLeave
  '/sickLeaves/{id}':
    summary: Specific operation on sickLeaves
    get:
      parameters:
        - examples:
            '1':
              value: 1
          name: id
          description: SickLeave name
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SickLeave'
              examples:
                '1':
                  value:
                    id: 1
                    accessmentId: 1
                    accidentId: 1
                    businessCode: A string
                    employeeId: A string
                    averageSalary: value
                    period: 1
                    status: A string
                    date: value
          description: sickLeaves with specified id
      operationId: GetSickLeaveById
      summary: Get SickLeave by id
      description: Get SickLeave by id
    patch:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SickLeave'
            examples:
              '1':
                value:
                  id: 1
                  accessmentId: 1
                  accidentId: 1
                  businessCode: A string
                  employeeId: A string
                  averageSalary: value
                  period: 1
                  status: A string
                  date: value
        required: true
      parameters:
        - examples:
            '1':
              value: 1
          name: id
          description: SickLeave name
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SickLeave'
              examples:
                '1':
                  value:
                    id: 1
                    accessmentId: 1
                    accidentId: 1
                    businessCode: A string
                    employeeId: A string
                    averageSalary: value
                    period: 1
                    status: A string
                    date: value
          description: Changed SickLeave
      operationId: PatchSickLeave
      summary: Patch existing SickLeave
    parameters:
      - name: id
        description: SickLeave id
        schema:
          type: integer
        in: path
        required: true
openapi: 3.1.0
info:
  title: SickLeave
  version: 0.0.1
  description: API definition of sickLeave
  contact:
    name: creator's name
    url: 'http://creator.homepage.address'
    email: creator@creator.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
components:
  schemas:
    SickLeave:
      title: Root Type for SickLeave
      description: The root of the SickLeave type's schema.
      type: object
      properties:
        id:
          description: id of this SickLeave
          type: integer
        accessmentId:
          description: accessmentId of this SickLeave
          type: integer
        accidentId:
          description: accidentId of this SickLeave
          type: integer
        businessCode:
          description: businessCode of this SickLeave
          type: string
        employeeId:
          description: employeeId of this SickLeave
          type: string
        averageSalary:
          description: averageSalary of this SickLeave
          type: string
        period:
          description: period of this SickLeave
          type: integer
        status:
          description: status of this SickLeave
          type: string
        date:
          description: date of this SickLeave
          type: string
      example:
        id: 1
        accessmentId: 1
        accidentId: 1
        businessCode: A string
        employeeId: A string
        averageSalary: value
        period: 1
        status: A string
        date: value
tags:
  - name: SickLeave
    description: SickLeave resource